Attempted shaders

+++ src/rand.cpp
@@ -5,5 +5,5 @@
 namespace specula::rand {
 std::random_device rd;
 std::mt19937 gen(rd());
-std::uniform_real_distribution<double> unif_rand_dist(0.0, 1.0);
+std::uniform_real_distribution<float> unif_rand_dist(0.0, 1.0);
 } // namespace specula::rand
